requirements.txt
Django==4.2.1
django-crontab==0.7.1
djangorestframework==3.14.0